Renamed axiosConfig to NetworkService
